Mac: Support mouse events for controls in a CustomCell
